added main, print config and assignment methods. Code executes. Problem with the cost function

+++ src/mcp/Task.java
@@ -22,7 +22,7 @@ private Integer priority;
     	this.WCET = WCET;
     	this.deadline = d;
         this.period = p;
-        this.priority = 1/period; // added priority calculation here (shortest period -> higher priority)
+        this.priority = Math.toIntExact(1/period); // added priority calculation here (shortest period -> higher priority)
     }
     
     public Long getDeadline() {
